safari-600.7-branch/LayoutTests/http/tests/websocket/tests/hybi/close-reason-too-long-expected.txt

CONSOLE MESSAGE: WebSocket close message is too long.
Checks whether SyntaxError is thrown when attempting to send too long reason.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Connected.
PASS SyntaxError was thrown.
PASS successfullyParsed is true

TEST COMPLETE
